refactor(task) : move function to new file
- manage function based on responsibility of function
